requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
WeasyPrint==61.2
pydyf==0.10.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
